# malloc []byte in cgo outside go heap