--- a/hfile/load-file.go
+++ b/hfile/load-file.go
@@ -5,9 +5,33 @@ import (
 	"io"
 	"log"
 	"os"
+	"reflect"
+	"unsafe"
 
 	"github.com/edsrzf/mmap-go"
 )
+
+/*
+#include<stdlib.h>
+*/
+import "C"
+
+/*
+Allocate a []byte outside the control of the garbage collector.
+
+This memory will never be freed -- don't use this unless you are sure that is what you want.
+
+Putting gigs and gigs of static, long-lived data on the gc's managed heap has the potential to
+throw off any heuristics which to use the total size of the heap (e.g. maintain some % free space).
+*/
+func offheapMalloc(size int) []byte {
+	hdr := reflect.SliceHeader{
+		Data: uintptr(unsafe.Pointer(C.malloc(C.size_t(size)))),
+		Len:  size,
+		Cap:  size,
+	}
+	return *(*[]byte)(unsafe.Pointer(&hdr))
+}
 
 func loadFile(name, path string, method LoadMethod) ([]byte, error) {
 	f, err := os.OpenFile(path, os.O_RDONLY, 0)
@@ -49,7 +73,7 @@ func loadFile(name, path string, method LoadMethod) ([]byte, error) {
 		defer f.Close()
 
 		log.Printf("[Reader.NewReader] Reading in %s (%.02fmb)...\n", name, sizeMb)
-		data := make([]byte, fi.Size())
+		data := offheapMalloc(int(fi.Size()))
 		if _, err := io.ReadFull(f, data); err != nil {
 			log.Printf("[Reader.NewReader] Error reading in %s: %s\n", name, err.Error())
 			return nil, err