Putting gigs and gigs of static, long-lived data on the gc's managed heap has the potential to throw off any heuristics which use the total size of the heap (e.g. maintain some % free space).

Benchmarks suggest no major change, though there's a tiny increase in alloc's for http (?):

Before:
```
    BenchmarkTBinaryHTTP-8          138425 ns/op     38924 B/op      116 allocs/op
    BenchmarkTCompactHTTP-8         126185 ns/op     39021 B/op      120 allocs/op
    BenchmarkTBinaryRaw-8            69787 ns/op      1327 B/op       42 allocs/op
    BenchmarkTCompactRaw-8           68698 ns/op      1326 B/op       42 allocs/op
    BenchmarkConcurrentHttp-8        41319 ns/op     41862 B/op      289 allocs/op
    BenchmarkConcurrentRaw-8         28322 ns/op      4251 B/op      215 allocs/op
    BenchmarkHandlerUncompressed-8   13948 ns/op       552 B/op       13 allocs/op
    BenchmarkHandlerCompressed-8     24396 ns/op       631 B/op       18 allocs/op
    BenchmarkHandlerMapped-8         24295 ns/op       632 B/op       18 allocs/op
```
After:
```
    BenchmarkTBinaryHTTP-8          124691 ns/op     39045 B/op      117 allocs/op
    BenchmarkTCompactHTTP-8         123113 ns/op     39132 B/op      121 allocs/op
    BenchmarkTBinaryRaw-8            65759 ns/op      1327 B/op       42 allocs/op
    BenchmarkTCompactRaw-8           66503 ns/op      1326 B/op       42 allocs/op
    BenchmarkConcurrentHttp-8        57054 ns/op     42019 B/op      290 allocs/op
    BenchmarkConcurrentRaw-8         29845 ns/op      4252 B/op      215 allocs/op
    BenchmarkHandlerUncompressed-8   14743 ns/op       552 B/op       13 allocs/op
    BenchmarkHandlerCompressed-8     25670 ns/op       631 B/op       18 allocs/op
    BenchmarkHandlerMapped-8         24559 ns/op       632 B/op       18 allocs/op
```